[frontend] StixNestedRefRelationshipEditionOverview in tsx
