+++ opencti-platform/opencti-front/src/private/components/common/stix_nested_ref_relationships/StixNestedRefRelationshipEdition.tsx
@@ -1,34 +1,28 @@
 import React, { FunctionComponent } from 'react';
-import { graphql } from 'react-relay';
 import Drawer from '@mui/material/Drawer';
 import { useTheme } from '@mui/styles';
-import { QueryRenderer } from '../../../../relay/environment';
-import StixNestedRefRelationshipEditionOverview from './StixNestedRefRelationshipEditionOverview';
+import { StixNestedRefRelationshipEditionQuery } from '@components/common/stix_nested_ref_relationships/__generated__/StixNestedRefRelationshipEditionQuery.graphql';
+import StixNestedRefRelationshipEditionOverview, { stixNestedRefRelationshipEditionQuery } from './StixNestedRefRelationshipEditionOverview';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
 import type { Theme } from '../../../../components/Theme';
-
-const stixNestedRefRelationshipEditionQuery = graphql`
-  query StixNestedRefRelationshipEditionQuery($id: String!) {
-    stixRefRelationship(id: $id) {
-      ...StixNestedRefRelationshipEditionOverview_stixRefRelationship
-    }
-  }
-`;
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
 
 interface StixNestedRefRelationshipEditionProps {
   stixNestedRefRelationshipId: string,
   open: boolean,
   handleClose?: () => void,
-  handleDelete?: () => boolean,
 }
 
 const StixNestedRefRelationshipEdition: FunctionComponent<StixNestedRefRelationshipEditionProps> = ({
   stixNestedRefRelationshipId,
   open,
   handleClose,
-  handleDelete,
 }) => {
   const theme = useTheme<Theme>();
+  const queryRef = useQueryLoading<StixNestedRefRelationshipEditionQuery>(
+    stixNestedRefRelationshipEditionQuery,
+    { id: stixNestedRefRelationshipId },
+  );
   return (
     <Drawer
       open={open}
@@ -48,29 +42,15 @@ const StixNestedRefRelationshipEdition: FunctionComponent<StixNestedRefRelations
       }}
       onClose={handleClose}
     >
-      {stixNestedRefRelationshipId ? (
-        <QueryRenderer
-          query={stixNestedRefRelationshipEditionQuery}
-          variables={{ id: stixNestedRefRelationshipId }}
-          render={({ props }) => {
-            if (props) {
-              return (
-                <StixNestedRefRelationshipEditionOverview
-                  stixRefRelationship={props.stixRefRelationship}
-                  handleClose={handleClose}
-                  handleDelete={
-                      typeof handleDelete === 'function'
-                        ? handleDelete
-                        : null
-                    }
-                />
-              );
-            }
-            return <Loader variant={LoaderVariant.inline} />;
-          }}
-        />
-      ) : (
-        <div> &nbsp; </div>
+      {queryRef && (
+        <React.Suspense
+          fallback={<Loader variant={LoaderVariant.inline} />}
+        >
+          <StixNestedRefRelationshipEditionOverview
+            queryRef={queryRef}
+            handleClose={handleClose}
+          />
+        </React.Suspense>
       )}
     </Drawer>
   );
